+++ src/core.js
@@ -233,11 +233,11 @@ function _buildSquel(flavour = null) {
      * Sanitize given expression.
      */
     _sanitizeExpression (expr) {
-      // If it's not an Expression builder instance
-      if (!(expr instanceof cls.Expression)) {
+      // If it's not a base builder instance
+      if (!(expr instanceof cls.BaseBuilder)) {
         // It must then be a string
         if (typeof expr !== "string") {
-          throw new Error("expression must be a string or Expression instance");
+          throw new Error("expression must be a string or builder instance");
         }
       }
 
@@ -274,7 +274,7 @@ function _buildSquel(flavour = null) {
         return item;
       }
 
-      throw new Error("must be a BaseBuilder instance");
+      throw new Error("must be a builder instance");
     }
 
 
@@ -283,7 +283,7 @@ function _buildSquel(flavour = null) {
         try {
           item = this._sanitizeBaseBuilder(item);
         } catch (e) {
-          throw new Error("table name must be a string or a query builder");
+          throw new Error("table name must be a string or a builder");
         }
       } else {
         item = this._sanitizeName(item, 'table');